backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
PyJWT
bcrypt==4.0.1
passlib
python-dotenv
